refactor: move logging into a directory
should tidy things up a bit
